add example springboot

+++ demo/src/main/java/com/arrizsoft/demo/dao/EmployeeDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.arrizsoft.demo.entity.Employee;
 
-public interface EmployeeDao extends PagingAndSortingRepository<Employee, Integer> {
+public interface EmployeeDao extends PagingAndSortingRepository<Employee, Long> {
 
 }
